Mixin-ify more view functionality
