+++ packages/ember-views/lib/mixins/legacy_view_support.js
@@ -1,0 +1,77 @@
+import Ember from 'ember-metal/core';
+import { Mixin } from "ember-metal/mixin";
+import { get } from "ember-metal/property_get";
+
+var LegacyViewSupport = Mixin.create({
+  mutateChildViews: function(callback) {
+    var childViews = this._childViews;
+    var idx = childViews.length;
+    var view;
+
+    while (--idx >= 0) {
+      view = childViews[idx];
+      callback(this, view, idx);
+    }
+
+    return this;
+  },
+
+  /**
+    Removes all children from the `parentView`.
+
+    @method removeAllChildren
+    @return {Ember.View} receiver
+  */
+  removeAllChildren: function() {
+    return this.mutateChildViews(function(parentView, view) {
+      parentView.removeChild(view);
+    });
+  },
+
+  destroyAllChildren: function() {
+    return this.mutateChildViews(function(parentView, view) {
+      view.destroy();
+    });
+  },
+
+  /**
+    Return the nearest ancestor whose parent is an instance of
+    `klass`.
+
+    @method nearestChildOf
+    @param {Class} klass Subclass of Ember.View (or Ember.View itself)
+    @return Ember.View
+    @deprecated
+  */
+  nearestChildOf: function(klass) {
+    Ember.deprecate("nearestChildOf has been deprecated.");
+
+    var view = get(this, 'parentView');
+
+    while (view) {
+      if (get(view, 'parentView') instanceof klass) { return view; }
+      view = get(view, 'parentView');
+    }
+  },
+
+  /**
+    Return the nearest ancestor that is an instance of the provided
+    class.
+
+    @method nearestInstanceOf
+    @param {Class} klass Subclass of Ember.View (or Ember.View itself)
+    @return Ember.View
+    @deprecated
+  */
+  nearestInstanceOf: function(klass) {
+    Ember.deprecate("nearestInstanceOf is deprecated and will be removed from future releases. Use nearestOfType.");
+    var view = get(this, 'parentView');
+
+    while (view) {
+      if (view instanceof klass) { return view; }
+      view = get(view, 'parentView');
+    }
+  }
+});
+
+export default LegacyViewSupport;
